+++ procedural/noise_heightfield.cpp
@@ -4,13 +4,13 @@
 #include "../core/noise.h"
 
 // Noise functions
-float perlinNoise(Vector p, int octaves, float lacunarity, float persistence, float scale){
-	float sum = 0.0f;
-	float frequency = scale;
-	float amplitude = 1.0f;
+double perlinNoise(Vector p, int octaves, double lacunarity, double persistence, double scale){
+	double sum = 0.0f;
+	double frequency = scale;
+	double amplitude = 1.0f;
 
 	for(int i = 0; i < octaves; i++) {
-		float n = noise2D(p.x * frequency, p.y * frequency);
+		double n = noise2D(p.x * frequency, p.y * frequency);
 		sum += n * amplitude;
 		frequency *= lacunarity;
 		amplitude *= persistence;
@@ -19,29 +19,29 @@ float perlinNoise(Vector p, int octaves, float lacunarity, float persistence, fl
 	return sum;
 }
 
-float billowNoise(Vector p, int octaves, float lacunarity, float gain, float scale){
+double billowNoise(Vector p, int octaves, double lacunarity, double gain, double scale){
 	return fabs(perlinNoise(p, octaves, lacunarity, gain, scale));
 }
 
-float ridgeNoise(Vector p, int octaves, float lacunarity, float gain, float scale){
+double ridgeNoise(Vector p, int octaves, double lacunarity, double gain, double scale){
 	return 1.0f - billowNoise(p, octaves, lacunarity, gain, scale);
 }
 
-float warpedNoise(Vector warp, float warpScale, Vector p, int octaves, float lacunarity, float persistence, float scale){
+double warpedNoise(Vector warp, double warpScale, Vector p, int octaves, double lacunarity, double persistence, double scale){
     Vector offset = Vector(perlinNoise(p, octaves, lacunarity, persistence, scale), 
             perlinNoise(p + warp, octaves, lacunarity, persistence, scale));
 
     return perlinNoise(p + offset * warpScale, octaves, lacunarity, persistence, scale);
 }
 
-void generateHeightfieldFromNoise(float** heightfield, int size, int octaves, float lacunarity, float persistence, float scale, float height){
-    float min = 0.0f;
-    float max = 0.0f;
+void generateHeightfieldFromNoise(double** heightfield, int size, int octaves, double lacunarity, double persistence, double scale, double height){
+    double min = 0.0f;
+    double max = 0.0f;
 
     for (int z = 0; z < size; z++){
         for (int x = 0; x < size; x++){
-            //float elevation = perlinNoise(Vector(x / (float)size, z / (float)size), octaves, lacunarity, persistence, scale);
-            float elevation = warpedNoise(Vector(5.2f, 1.3f), 0.0f, Vector(x / (float)size, z / (float)size), octaves, lacunarity, persistence, scale);
+            //double elevation = perlinNoise(Vector(x / (double)size, z / (double)size), octaves, lacunarity, persistence, scale);
+            double elevation = warpedNoise(Vector(5.2f, 1.3f), 0.0f, Vector(x / (double)size, z / (double)size), octaves, lacunarity, persistence, scale);
             heightfield[z][x] = elevation;
 
             if (elevation < min){
